requirements.txt
Django>=4.2.0,<5.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=2023.2
Pillow>=10.0.0
celery==5.3.6
redis==5.0.1
django-celery-results==2.5.1
whitenoise==6.6.0 